+++ src/Symfony/Component/WebLink/EventListener/AddLinkHeaderListener.php
@@ -16,12 +16,14 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\WebLink\HttpHeaderSerializer;
+use Symfony\Component\WebLink\LinkTemplateHeaderSerializer;
 
 // Help opcache.preload discover always-needed symbols
 class_exists(HttpHeaderSerializer::class);
+class_exists(LinkTemplateHeaderSerializer::class);
 
 /**
- * Adds the Link HTTP header to the response.
+ * Adds the Link and Link-Template HTTP headers to the response.
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
  *
@@ -31,6 +33,7 @@ class AddLinkHeaderListener implements EventSubscriberInterface
 {
     public function __construct(
         private readonly HttpHeaderSerializer $serializer = new HttpHeaderSerializer(),
+        private readonly LinkTemplateHeaderSerializer $templateSerializer = new LinkTemplateHeaderSerializer(),
     ) {
     }
 
@@ -45,7 +48,15 @@ class AddLinkHeaderListener implements EventSubscriberInterface
             return;
         }
 
-        $event->getResponse()->headers->set('Link', $this->serializer->serialize($links), false);
+        $headers = $event->getResponse()->headers;
+
+        if (null !== $header = $this->serializer->serialize($links)) {
+            $headers->set('Link', $header, false);
+        }
+
+        if (null !== $header = $this->templateSerializer->serialize($links)) {
+            $headers->set('Link-Template', $header, false);
+        }
     }
 
     public static function getSubscribedEvents(): array