+++ src/Symfony/Component/WebLink/HttpHeaderParser.php
@@ -16,7 +16,7 @@ use Psr\Link\EvolvableLinkProviderInterface;
 /**
  * Parse a list of HTTP Link headers into a list of Link instances.
  *
- * @see https://tools.ietf.org/html/rfc5988
+ * @see https://www.rfc-editor.org/rfc/rfc8288.html
  *
  * @author Jérôme Tamarelle <jerome@tamarelle.net>
  */